What is happening in the "Auction.sol" Smart Contract ?
Anyone can register in the Auction but a maximum of 4 persons can take part and there are 3 items in the Auction. The beneficiary(manager/the person who deployed the smart contract) has the power to reveal winners, no other person can choose a winner. The bid is an amount to be paid, a person can bid on any of the items. "getPersonDetails" is a calling function to get the details of a particular person by putting the address of the person in the box provided. In the end, a winner is declared by the beneficiary.